requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
azure-core==1.30.1
azure-identity==1.16.0
black==24.4.2
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
dataclasses-json==0.6.5
distro==1.9.0
Flask==3.0.3
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.50
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langsmith==0.1.54
MarkupSafe==2.1.5
marshmallow==3.21.2
msal==1.28.0
msal-extensions==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.25.1
orjson==3.10.3
packaging==23.2
pathspec==0.12.1
pip==24.0
platformdirs==4.2.1
portalocker==2.8.2
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyodbc==5.1.0
pywin32==306
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.4
typing_extensions==4.11.0
typing-inspect==0.9.0
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
